refactor: Test코드에서만 필요한 MoveNumberForTest 클래스를 innerClass로 전환
