--- src/test/java/racingcar/domain/MoveNumberForTest.java
@@ -1,14 +1,0 @@
-package racingcar.domain;
-
-public class MoveNumberForTest extends MoveNumber {
-
-    private final int moveNumber;
-
-    public MoveNumberForTest(int moveNumber) {
-        this.moveNumber = moveNumber;
-    }
-
-    public int getMoveNumber() {
-        return this.moveNumber;
-    }
-}
